refactor(docs): extract _rewrite_sitemap_index into a Sphinx extension
Moves _rewrite_sitemap_index (event: build-finished) from conf.py into
docs/source/_extensions/sitemap_canonical.py. Rewrites the homepage URL
in the emitted sitemap.xml to the bare-host canonical form, keeping
sphinx-sitemap output aligned with the canonical override set in
_inject_seo_context.
Helpers moved alongside: none (self-contained). conf.py shrinks from 4485
to 4426 LOC.

Part of C1 in docs_pipeline_architecture_review.md
